Move Schema to type.h, remove Equals with shared_ptr function
