+++ cpp/src/arrow/ipc/json-internal.cc
@@ -33,7 +33,6 @@
 #include "arrow/array.h"
 #include "arrow/builder.h"
 #include "arrow/memory_pool.h"
-#include "arrow/schema.h"
 #include "arrow/status.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
