+++ cpp/src/arrow/ipc/reader.cc
@@ -30,9 +30,9 @@
 #include "arrow/ipc/Message_generated.h"
 #include "arrow/ipc/metadata.h"
 #include "arrow/ipc/util.h"
-#include "arrow/schema.h"
 #include "arrow/status.h"
 #include "arrow/table.h"
+#include "arrow/type.h"
 #include "arrow/util/logging.h"
 
 namespace arrow {
