+++ cpp/src/arrow/ipc/writer.cc
@@ -32,7 +32,6 @@
 #include "arrow/ipc/util.h"
 #include "arrow/loader.h"
 #include "arrow/memory_pool.h"
-#include "arrow/schema.h"
 #include "arrow/status.h"
 #include "arrow/table.h"
 #include "arrow/type.h"
