+++ cpp/src/arrow/python/pandas-test.cc
@@ -25,7 +25,6 @@
 #include "arrow/array.h"
 #include "arrow/builder.h"
 #include "arrow/python/pandas_convert.h"
-#include "arrow/schema.h"
 #include "arrow/table.h"
 #include "arrow/test-util.h"
 #include "arrow/type.h"
